refactor: move parameters to _parameters
