+++ src/awkward/_broadcasting.py
@@ -12,6 +12,7 @@ from awkward._behavior import find_custom_broadcast
 from awkward._nplikes.numpy import Numpy
 from awkward._nplikes.numpylike import NumpyMetadata
 from awkward._nplikes.shape import unknown_length
+from awkward._parameters import _parameters_intersect, _parameters_is_empty
 from awkward._typing import Any, Callable, Dict, List, TypeAlias, Union
 from awkward._util import unset
 from awkward.contents.bitmaskedarray import BitMaskedArray
@@ -278,14 +279,14 @@ def intersection_parameters_factory(
     # If we encounter None-parameters, then we stop early
     # as there can be no intersection.
     for parameters in input_parameters:
-        if ak.forms.form._parameters_is_empty(parameters):
+        if _parameters_is_empty(parameters):
             break
         else:
             parameters_to_intersect.append(parameters)
     # Otherwise, build the intersected parameter dict
     else:
         intersected_parameters = functools.reduce(
-            ak.forms.form._parameters_intersect, parameters_to_intersect
+            _parameters_intersect, parameters_to_intersect
         )
 
     def apply(n_outputs: int) -> list[dict[str, Any] | None]:
