+++ src/awkward/forms/recordform.py
@@ -4,10 +4,11 @@ from collections.abc import Iterable
 
 import awkward as ak
 from awkward._behavior import find_typestr
+from awkward._parameters import _type_parameters_equal
 from awkward._regularize import is_integer
 from awkward._typing import final
 from awkward._util import unset
-from awkward.forms.form import Form, _type_parameters_equal
+from awkward.forms.form import Form
 
 
 @final
