+++ src/awkward/forms/unmaskedform.py
@@ -1,9 +1,10 @@
 # BSD 3-Clause License; see https://github.com/scikit-hep/awkward-1.0/blob/main/LICENSE
 import awkward as ak
 from awkward._behavior import find_typestr
+from awkward._parameters import _parameters_union, _type_parameters_equal
 from awkward._typing import final
 from awkward._util import unset
-from awkward.forms.form import Form, _type_parameters_equal
+from awkward.forms.form import Form
 
 
 @final
@@ -57,15 +58,11 @@ class UnmaskedForm(Form):
         if content.is_union:
             return content.copy(
                 contents=[cls.simplified(x) for x in content.contents],
-                parameters=ak.forms.form._parameters_union(
-                    content._parameters, parameters
-                ),
+                parameters=_parameters_union(content._parameters, parameters),
             )
         elif content.is_indexed or content.is_option:
             return content.copy(
-                parameters=ak.forms.form._parameters_union(
-                    content._parameters, parameters
-                )
+                parameters=_parameters_union(content._parameters, parameters)
             )
         else:
             return cls(content, parameters=parameters, form_key=form_key)
