+++ src/awkward/types/numpytype.py
@@ -5,8 +5,8 @@ import re
 
 import awkward as ak
 from awkward._nplikes.numpylike import NumpyMetadata
+from awkward._parameters import _type_parameters_equal
 from awkward._typing import final
-from awkward.forms.form import _type_parameters_equal
 from awkward.types.type import Type
 
 np = NumpyMetadata.instance()
