+++ src/awkward/types/uniontype.py
@@ -3,8 +3,8 @@
 from collections.abc import Iterable
 
 import awkward as ak
+from awkward._parameters import _type_parameters_equal
 from awkward._typing import final
-from awkward.forms.form import _type_parameters_equal
 from awkward.types.type import Type
